requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core==1.31.0
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pytest==8.3.3
requests==2.32.3
six==1.16.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3